+++ src/app.d.ts
@@ -1,10 +1,12 @@
 // See https://svelte.dev/docs/kit/types#app.d.ts
 // for information about these interfaces
+import type { SupabaseClient } from '@supabase/supabase-js';
+
 declare global {
 	namespace App {
 		interface Locals {
+			supabase: SupabaseClient;
 			user?: { id: string; name: string; email: string; image?: string | null };
-			session?: { id: string; userId: string; expiresAt: string };
 		}
 
 		// interface Error {}